+++ tests/Models/InventoryTest.lua
@@ -4,11 +4,35 @@ TestInventory = BaseTestClass:new()
         local instance = __:new('Inventory')
 
         lu.assertNotNil(instance)
+        lu.assertTrue(instance.outdated)
+        lu.assertEquals({}, instance.containers)
+    end
+
+    -- @covers Inventory:flagOutdated()
+    function TestInventory:testFlagOutdated()
+        local instance = __:new('Inventory')
+
+        local containerMock = __:new('Container')
+        containerMock.flagOutdated = function ()
+            containerMock.flagOutdatedInvoked = true
+            return containerMock
+        end
+
+        instance.containers = { containerMock }
+
+        instance.outdated = false
+
+        local result = instance:flagOutdated()
+
+        lu.assertTrue(instance.outdated)
+        lu.assertTrue(containerMock.flagOutdatedInvoked)
+        lu.assertEquals(instance, result)
     end
 
     -- @covers Inventory:getItems()
     function TestInventory:testGetItems()
         local instance = __:new('Inventory')
+        instance.maybeMapContainers = function () instance.maybeMapContainersInvoked = true end
 
         local containerA, containerB = __:new('Container'), __:new('Container')
 
@@ -20,16 +44,19 @@ TestInventory = BaseTestClass:new()
         local result = instance:getItems()
 
         lu.assertEquals({ 'itemA1', 'itemA2', 'itemB1', 'itemB2' }, result)
+        lu.assertTrue(instance.maybeMapContainersInvoked)
     end
 
     -- @covers Inventory:hasItem()
     function TestInventory:testHasItem()
         local function execution(containers, expectedOutput)
             local instance = __:new('Inventory')
+            instance.maybeMapContainers = function () instance.maybeMapContainersInvoked = true end
 
             instance.containers = containers
 
             lu.assertEquals(expectedOutput, instance:hasItem())
+            lu.assertTrue(instance.maybeMapContainersInvoked)
         end
 
         local containerHasItem = { hasItem = function () return true end }
@@ -51,6 +78,7 @@ TestInventory = BaseTestClass:new()
     -- @covers Inventory:mapContainers()
     function TestInventory:testMapContainers()
         local instance = __:new('Inventory')
+        instance.outdated = true
 
         __.arr:maybeInitialize(_G, 'Enum.BagIndex', { TestBag = 1 })
 
@@ -65,6 +93,7 @@ TestInventory = BaseTestClass:new()
         local result = instance:mapContainers()
 
         lu.assertTrue(bagMock.mapItemsInvoked)
+        lu.assertFalse(instance.outdated)
         lu.assertEquals({ bagMock }, instance.containers)
         lu.assertEquals(instance, result)
 
@@ -80,9 +109,34 @@ TestInventory = BaseTestClass:new()
         lu.assertIsNil(instance.bags)
     end
 
+    -- @covers Inventory:maybeMapContainers()
+    function TestInventory:testMaybeMapContainers()
+        local function execution(outdated, shouldCallMapContainers)
+            local instance = __:new('Inventory')
+            instance.mapContainersInvoked = false
+            instance.outdated = outdated
+
+            instance.mapContainers = function ()
+                instance.mapContainersInvoked = true
+                return instance
+            end
+
+            instance:maybeMapContainers()
+
+            lu.assertEquals(shouldCallMapContainers, instance.mapContainersInvoked)
+        end
+
+        -- outdated
+        execution(true, true)
+
+        -- not outdated
+        execution(false, false)
+    end
+
     -- @covers Inventory:refresh()
     function TestInventory:testRefresh()
         local instance = __:new('Inventory')
+        instance.maybeMapContainers = function () instance.maybeMapContainersInvoked = true end
 
         local containerMock = __:new('Container')
         containerMock.refresh = function ()
@@ -95,6 +149,7 @@ TestInventory = BaseTestClass:new()
         local result = instance:refresh()
 
         lu.assertTrue(containerMock.refreshInvoked)
+        lu.assertTrue(instance.maybeMapContainersInvoked)
         lu.assertEquals(instance, result)
     end
 -- end of TestInventory